requirements.txt
langchain==0.1.13
langchain-community==0.0.29
openai==1.14.2
streamlit==1.30.0
python-dotenv==1.0.1
requests==2.31.0
